+++ Assets/Scripts/SpellEditor/ComponentPanel/ProjectilePanel.cs
@@ -47,10 +47,10 @@ namespace SpellEditor.ComponentPanel
             
             ProjectileSpell projSpell = new ProjectileSpell
             {
-                damages = damage.text != "" ? Int32.Parse(damage.text) : 0,
-                duration = duration.text != "" ? Int32.Parse(duration.text) : 0,
-                speed = speed.text != "" ? Int32.Parse(speed.text) : 1,
-                damageMultiplierOnDistance = multiplier.text != "" ? Int32.Parse(multiplier.text) : 0,
+                damages = damage.text != "" ? float.Parse(damage.text) : 0,
+                duration = duration.text != "" ? float.Parse(duration.text) : 0,
+                speed = speed.text != "" ? float.Parse(speed.text) : 1,
+                damageMultiplierOnDistance = multiplier.text != "" ? float.Parse(multiplier.text) : 0,
                 idPoolObject = Int32.Parse(idPoolObject.text),
                 passingThroughEntity = passingThrough.isOn,
                 linkedSpellOnEnable = linkedSpellOnEnable.value != 0 ? ListCreatedElement.SpellComponents[linkedSpellOnEnable.options[linkedSpellOnEnable.value].text] : null,
